read and write gzipped SBML
Not super important, as libSBML can usually handle compressed model
file. However, in the homebrew/brewsci build of SBML, the gzip-support
is disables due to a conflict of libSBML and zlib:

https://github.com/brewsci/homebrew-bio/pull/1672

The change here, used R's build-in gz-file in/out support to read/write
gzipped SBML files.

Note: It is a little less efficient than the libSBML gz-support. When
zlib-connection in homebrew's libsbml is back, the lines added in this
commit could be removed again.