Objective direction as slot in ModelOrg object
I have decided to make the objective direction (maximize/minimize) a
slot of each instance of "ModelOrg". Before, it was a global option set
in COBRA_SETTINGS("OPT_DIRECTION"). Now, models can have independent
directions for their objective directions.

This is also to comply with LIBSBML, where the direction is art of each
objective object.