Enforce global maximum for absolute flux values
A global absolute flux value can be set via
`COBRAR_SETTING("MAXIMUM",<value>")`. Yet, this setting was not enforced
in previous versions. Now: Before every flux predictions, it is checked
if some lower or upper bounds have an absolute value higher than the
global maximum. If yes, this bound are reset to the global MAX/MIN and a
warning is returned.

Fixes #7.