Better approach to ensure globally set maximum flux
COBRA_SETTINGS("MAXIMUM") hold the value for the absolute maximum flux
rate per reaction. In the previos version, the flux bounds were changes
in the ModelOrg object before initializing the LP-problem, which causes
a memory copy of the ModelOrg object. Now, the flux bounds are adjusted
direktly when defining the LP-problem, without changes in the ModelOrg
object.

Minor changes:
- roxygen2 version from 7.3.2 to 7.3.3
- Autoconf version update to create `configure` script
- Last name changes of contributor
- cobrar version increase to v0.2.1