Generalized identification of exchange reactions
Before: Exchange reactions were identfied via the "EX_" prefix in the
reaction ID. However, not all models follow this convention.

Now: Exchange reactions are identified by their stoichiometry. In brief:
Exchange reactions are those unbalanced reactions, that have only one
metabolite participating and where the metabolite occures either on the
LHS or the RHS of the equation, but not both.

Closes #25